[Feat] 화면 전환 구현 및 최근모델 api 작업

+++ lib/model/GetModelrecent.dart
@@ -29,7 +29,7 @@ class Data {
   int? modelId;
   String? imageUrl;
   String? modelName;
-  int? aiRate;
+  double? aiRate;
   String? job;
 
   Data({this.modelId, this.imageUrl, this.modelName, this.aiRate, this.job});
